ENH Miliseconds in mysql and postgresql

+++ wopmars/framework/database/tables/Execution.py
@@ -1,8 +1,7 @@
-from sqlalchemy import Column, Integer, DateTime, Float, String
+from sqlalchemy import Column, Integer, DateTime, Float, String, BigInteger
 from sqlalchemy.orm import relationship
 
 from wopmars.framework.database.Base import Base
-
 
 class Execution(Base):
     """
@@ -12,8 +11,8 @@ class Execution(Base):
     The table ``wom_execution`` contains the following fields:
 
     - id: INTEGER: primary key - auto increment - arbitrary iD
-    - started_at: INTEGER - unix time at wich the execution started
-    - finished_at: INTEGER - unix time at wich the execution finished
+    - started_at: BigInteger - unix time at wich the execution started
+    - finished_at: BigInteger - unix time at wich the execution finished
     - time: FLOAT - the total time of execution
     - status: VARCHAR(255) - the final status of the Execution. it can be:
 
@@ -25,8 +24,8 @@ class Execution(Base):
     __tablename__ = "wom_execution"
 
     id = Column(Integer, primary_key=True, autoincrement=True)
-    started_at = Column(Integer, nullable=True)
-    finished_at = Column(Integer, nullable=True)
+    started_at = Column(BigInteger, nullable=True)
+    finished_at = Column(BigInteger, nullable=True)
     time = Column(Float, nullable=True)
     status = Column(String(255), nullable=True)
